refactor: add static router
