+++ task3/src/routes/auth.router.js
@@ -1,5 +1,5 @@
 const express = require("express");
-var router = express.Router();
+let router = express.Router();
 
 const pathCreator = require("../../public/js/pathCreator");
 const hasher = require('../../public/js/hasher');
